feat: Banner 분리
