+++ src/components/SurveyDetail/index.tsx
@@ -1,3 +1,3 @@
 export { default as Tab } from './Tab/Tab';
 export { default as Thumbnail } from './Thumbnail/Thumbnail';
-export { default as Banner } from './Banner/Banner';
+export { default as SurveyDetailBanner } from './SurveyDetailBanner/SurveyDetailBanner';
